config be & esp : mqtt commmunication
kurang send to ws

+++ Backend/dto/sensor_raw.go
@@ -1,7 +1,7 @@
 package dto
 
 type SensorData struct {
-	Temperature  float32 `json:"temperature"`
-	Humidity     float32 `json:"humidity"`
-	SoilMoisture float32 `json:"soil_moisture"`
+	Temperature  float64 `json:"temperature"`
+	Humidity     float64 `json:"humidity"`
+	SoilMoisture float64 `json:"soil_moisture"`
 }
